requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
autocommand==2.2.2
backports.tarfile==1.2.0
bcrypt==4.3.0
billiard==4.2.1
boto3==1.38.41
botocore==1.38.41
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.3
django-cacheops==7.2
django-ckeditor==6.7.3
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-js-asset==3.1.2
django-modeltranslation==0.19.11
django-redis==5.4.0
django-rosetta==0.10.1
django-silk==5.4.0
django-storages==1.14.6
django-unfold==0.42.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
funcy==2.0
gprof2dot==2025.4.14
h11==0.16.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.3.1
inflection==0.5.1
iniconfig==2.1.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jst-django-core==1.1.8
kombu==5.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
packaging==24.2
pillow==11.0.0
pip-chill==1.0.3
platformdirs==4.3.6
pluggy==1.6.0
polib==1.2.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rich==14.0.0
rpds-py==0.21.0
s3transfer==0.13.0
six==1.16.0
sqlparse==0.5.2
tomli==2.2.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.21.0
